per-type len()
Each type provides a len() function to return the, user-visible,
length of an object of that type. E.g. an array type returns the
number of elements in the array, a string the number of characters.
A default len() implementation returns 1 and only those types that
can be > 1 bother to override the default.