adding a repl (and re-formatting as I go)
Now when run without args ici enters a REPL-loop. It's not a real
REPL as it doesn't print anything (yet) but it does prompt for
input and parse code, report any errors and continue.

Getting an implicit print happening is another matter. It really
depends on the type of statement being parsed and requires more
hooks into that code.