vec: clean up functions to create vec objects
vec(bits, size, [fill]) -> creates a vec32 or vec64 according to the
value of bits (32 or 64) with size elements. Initialized to fill if
it is passed.

vec32() and vec64() now create objects initialized from their
arguments and take the place of literal forms.

vec32(1, 2, 3, 4, 5) returns a vec32 object of size 5 with the
obvious values.