Mention ipp in the interpreter configuration string iff enabled
Added the "(ipp)" to the config string to make it easier to
determine if IPP is enabled in a build.  This sort of thing
really should be part of a "feature set" accessible to code
via a map holding various bits of meta-data (an old idea
never put in) but suffices for the time being.